cv_music/requirements.txt
fastapi==0.87.0
mediapipe==0.9.0
numpy==1.21.6
opencv_python==4.6.0.66
Pillow==9.3.0
python-multipart
tensorflow==2.11.0
uvicorn==0.20.0
